pyproject.toml
[project]
name = "streamlit-cropper-fix"
version = "0.4.1"
authors = [{name="Jia Yi", email="e0774928@u.nus.edu"}]
description = "A simple image cropper for Streamlit, with fixes added"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit >= 1.27.0",
    "Pillow >= 9.2.0",
    "numpy >= 1.23.5"
]

[project.urls]
"Homepage" = "https://github.com/ajiayi-debug/streamlit-cropper"
"Bug Tracker" = "https://github.com/ajiayi-debug/streamlit-cropper/issues"

[build-system]
requires = ["setuptools>=61.0",
            "streamlit >= 1.27.0",
            "Pillow >= 9.2.0",
            "numpy >= 1.23.5"
]
build-backend = "setuptools.build_meta"